requirements.txt
crewai
langchain-openai
python-dotenv
litellm
